+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "start": "node index.js -- --username=Ivan"
+    "start": "node --watch index.mjs -- --username=Ivan"
   },
   "author": "",
   "license": "ISC"
